requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
cffi==1.11.5
click==6.7
colorama==0.3.9
cryptography==2.3
easypysmb==1.4.3
flake8==3.5.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
libkeepass==0.3.0
lxml==4.2.3
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
Pillow==5.2.0
pluggy==0.7.1
py==1.5.4
pyasn1==0.4.4
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.6.4
pyflakes==1.6.0
PyJWT==1.6.4
pykeepass==2.8.2
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyotp==2.2.6
pysmb==1.1.25
pytest==3.7.0
python-dateutil==2.7.3
qrcode==6.0
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
